Add files via upload
Now it works again, 
We have created new classes SFaceEmbeddings, KNNClassifier, UnrecognitionCriterion, to decrease the length of the FaceRecognition Class that had reach an absurdly length of about 550 lines... (ok several lines are actually comments... but still... )

TODO: 
- address the performance issue ( use shared_memory instead of queue to transfer images between processes.. ) 
- explore the structure of the classification problem in the space of SFace features:   we use k-nearest-neigbours since the classification in the SFace features is supposed to be very easy ( in toy examples, they use a simple distance-based threshold  ). 
But we realize that often there is circonstances where the classification is absurd. Francois is classified as Audrey for instance with very high certainty even if the classification in the same video is correct most of the time.